moved cmd handle
